Try to use TypeScript for some source
