--- runtime/examples/01_ts.ts
@@ -1,5 +1,0 @@
-// Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
-// deno-lint-ignore-file
-
-const foo: string = "foo";
-globalThis.foo = foo;
